workaround for removing the units with degree for ringcharts
this works because ° is 2 bytes wide, so any temperature with 2 or 3 digits will
be reduced to its number.
If the temperature is single digit (<= 9°C) the degree sign will stay. If the
temperature is above 1000°R (282°C) the number will be wrong, but at that point
the cpu/gpu is probably already dead.

I think ideally the RingChart constructor should receive some number with unit
object, and decide from the number if it should display the unit or not.